third_party/WebKit/LayoutTests/fullscreen/full-screen-iframe-legacy-expected.txt

CONSOLE WARNING: line 1: 'HTMLVideoElement.webkitDisplayingFullscreen' is deprecated. Please use 'Document.fullscreenElement' instead.
CONSOLE WARNING: line 10: 'HTMLVideoElement.webkitEnterFullScreen()' is deprecated. Please use 'Element.requestFullscreen()' instead.
HTMLVideoElement.webkitEnterFullScreen() in an iframe requires the allowfullscreen attribute

EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
EVENT(webkitfullscreenerror)
END OF TEST
